Allow connecting to a local VM in testenv
Currently, the VM is specified by address field.
This address is passed as a -ComputerName to New-PSSession.

This PR adds a possibility to connect to local Hyper-V VM,
by specifying vmName instead of address.

The goal of this PR is to make easier to use testenv-conf.yaml
for testing locally.